+++ internal/database/migration/018_create_client_uris_table.go
@@ -8,14 +8,14 @@ func CreateClientURIsTable(db *gorm.DB) error {
 	sql := `
 -- CREATE TABLE
 CREATE TABLE IF NOT EXISTS client_uris (
-    client_uri_id   SERIAL PRIMARY KEY,
+    client_uri_id   BIGSERIAL PRIMARY KEY,
     client_uri_uuid UUID NOT NULL UNIQUE,
-    tenant_id            INTEGER NOT NULL,
-    client_id       INTEGER NOT NULL,
-    uri                  TEXT NOT NULL,
-    type                 VARCHAR(20) NOT NULL DEFAULT 'redirect-uri',
-    created_at           TIMESTAMPTZ DEFAULT now(),
-    updated_at           TIMESTAMPTZ DEFAULT now()
+    tenant_id       BIGINT NOT NULL,
+    client_id       BIGINT NOT NULL,
+    uri             TEXT NOT NULL,
+    type            VARCHAR(20) NOT NULL DEFAULT 'redirect-uri',
+    created_at      TIMESTAMPTZ DEFAULT now(),
+    updated_at      TIMESTAMPTZ DEFAULT now()
 );
 
 -- ADD CONSTRAINTS (safe)
@@ -46,18 +46,12 @@ BEGIN
 END$$;
 
 -- ADD INDEXES
-CREATE INDEX IF NOT EXISTS idx_client_uris_uuid 
-    ON client_uris (client_uri_uuid);
-CREATE INDEX IF NOT EXISTS idx_client_uris_tenant_id 
-    ON client_uris (tenant_id);
-CREATE INDEX IF NOT EXISTS idx_client_uris_client_id 
-    ON client_uris (client_id);
-CREATE INDEX IF NOT EXISTS idx_client_uris_uri 
-    ON client_uris (uri);
-CREATE INDEX IF NOT EXISTS idx_client_uris_type 
-    ON client_uris (type);
-CREATE INDEX IF NOT EXISTS idx_client_uris_client_id_type 
-    ON client_uris (client_id, type);
+CREATE INDEX IF NOT EXISTS idx_client_uris_uuid ON client_uris (client_uri_uuid);
+CREATE INDEX IF NOT EXISTS idx_client_uris_tenant_id ON client_uris (tenant_id);
+CREATE INDEX IF NOT EXISTS idx_client_uris_client_id ON client_uris (client_id);
+CREATE INDEX IF NOT EXISTS idx_client_uris_uri ON client_uris (uri);
+CREATE INDEX IF NOT EXISTS idx_client_uris_type ON client_uris (type);
+CREATE INDEX IF NOT EXISTS idx_client_uris_client_id_type ON client_uris (client_id, type);
 `
 	return db.Exec(sql).Error
 }
